+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/samples/SamplesAjaxController.java
@@ -2,17 +2,13 @@ package ca.corefacility.bioinformatics.irida.ria.web.samples;
 
 import java.io.IOException;
 import java.security.Principal;
-import java.util.*;
+import java.util.List;
+import java.util.Locale;
+import java.util.Set;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
-
-import ca.corefacility.bioinformatics.irida.exceptions.ConcatenateException;
-import ca.corefacility.bioinformatics.irida.exceptions.EntityExistsException;
-import ca.corefacility.bioinformatics.irida.exceptions.EntityNotFoundException;
-import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
-import ca.corefacility.bioinformatics.irida.ria.web.samples.dto.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -22,16 +18,17 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 
+import ca.corefacility.bioinformatics.irida.exceptions.ConcatenateException;
+import ca.corefacility.bioinformatics.irida.exceptions.EntityExistsException;
+import ca.corefacility.bioinformatics.irida.exceptions.EntityNotFoundException;
 import ca.corefacility.bioinformatics.irida.model.sample.Sample;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequenceFile;
-
+import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxErrorResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxSuccessResponse;
+import ca.corefacility.bioinformatics.irida.ria.web.samples.dto.*;
 import ca.corefacility.bioinformatics.irida.ria.web.services.UIAnalysesService;
-
-import ca.corefacility.bioinformatics.irida.ria.web.samples.dto.SampleDetails;
-
 import ca.corefacility.bioinformatics.irida.ria.web.services.UISampleService;
 
 /**
@@ -62,8 +59,7 @@ public class SamplesAjaxController {
 		try {
 			return ResponseEntity.ok(uiSampleService.uploadSequenceFiles(sampleId, request));
 		} catch (IOException e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-					.body(null);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
 		}
 	}
 
@@ -80,8 +76,7 @@ public class SamplesAjaxController {
 		try {
 			return ResponseEntity.ok(uiSampleService.uploadFast5Files(sampleId, request));
 		} catch (IOException e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-					.body(null);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
 		}
 	}
 
@@ -98,8 +93,7 @@ public class SamplesAjaxController {
 		try {
 			return ResponseEntity.ok(uiSampleService.uploadAssemblies(sampleId, request));
 		} catch (IOException e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-					.body(null);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
 		}
 	}
 
@@ -145,8 +139,7 @@ public class SamplesAjaxController {
 			for (ConstraintViolation a : e.getConstraintViolations()) {
 				constraintViolations += a.getMessage() + "\n";
 			}
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new AjaxErrorResponse(constraintViolations));
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new AjaxErrorResponse(constraintViolations));
 		}
 	}
 
@@ -165,17 +158,16 @@ public class SamplesAjaxController {
 			return ResponseEntity.ok(new AjaxSuccessResponse(
 					uiSampleService.updateDefaultSequencingObjectForSample(id, sequencingObjectId, locale)));
 		} catch (EntityNotFoundException e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new AjaxErrorResponse(e.getMessage()));
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new AjaxErrorResponse(e.getMessage()));
 		}
 	}
 
 	/**
 	 * Update the default genome assembly for the sample
 	 *
-	 * @param id                 {@link Long} identifier for the sample
+	 * @param id               {@link Long} identifier for the sample
 	 * @param genomeAssemblyId The genome assembly identifier
-	 * @param locale             {@link Locale} for the currently logged in user
+	 * @param locale           {@link Locale} for the currently logged in user
 	 * @return {@link ResponseEntity} explaining to the user the results of the update.
 	 */
 	@PutMapping(value = "/{id}/default-genome-assembly")
@@ -185,8 +177,7 @@ public class SamplesAjaxController {
 			return ResponseEntity.ok(new AjaxSuccessResponse(
 					uiSampleService.updateDefaultGenomeAssemblyForSample(id, genomeAssemblyId, locale)));
 		} catch (EntityNotFoundException e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new AjaxErrorResponse(e.getMessage()));
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new AjaxErrorResponse(e.getMessage()));
 		}
 	}
 
@@ -204,8 +195,7 @@ public class SamplesAjaxController {
 		try {
 			return ResponseEntity.ok(uiAnalysesService.getSampleAnalyses(sampleId, principal, locale));
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(null);
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
 
@@ -227,16 +217,17 @@ public class SamplesAjaxController {
 	 * Remove metadata field and entry from {@link Sample}
 	 *
 	 * @param projectId       The project identifier
-	 * @param metadataFieldId   The metadata field id
+	 * @param sampleId        The sample identifier
+	 * @param metadataFieldId The metadata field id
 	 * @param metadataEntryId The metadata entry identifier
 	 * @param locale          {@link Locale} for the currently logged in user
 	 * @return {@link ResponseEntity} explaining to the user the results of the deletion.
 	 */
 	@DeleteMapping(value = "/metadata")
-	public ResponseEntity<AjaxResponse> removeSampleMetadata(@RequestParam Long projectId,
+	public ResponseEntity<AjaxResponse> removeSampleMetadata(@RequestParam Long projectId, @RequestParam Long sampleId,
 			@RequestParam Long metadataFieldId, @RequestParam Long metadataEntryId, Locale locale) {
 		return ResponseEntity.ok(new AjaxSuccessResponse(
-				uiSampleService.removeSampleMetadata(projectId, metadataFieldId, metadataEntryId, locale)));
+				uiSampleService.removeSampleMetadata(projectId, sampleId, metadataFieldId, metadataEntryId, locale)));
 	}
 
 	/**
@@ -254,8 +245,7 @@ public class SamplesAjaxController {
 			return ResponseEntity.ok(new AjaxSuccessResponse(
 					uiSampleService.updateSampleMetadata(id, updateSampleMetadataRequest, locale)));
 		} catch (EntityExistsException e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new AjaxErrorResponse(e.getMessage()));
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new AjaxErrorResponse(e.getMessage()));
 		}
 	}
 
@@ -273,7 +263,6 @@ public class SamplesAjaxController {
 	}
 
 	/**
-
 	 * Get updated sample sequencing objects for given sequencing object ids
 	 *
 	 * @param id                  Identifier for a sample
@@ -341,8 +330,7 @@ public class SamplesAjaxController {
 			return ResponseEntity.ok(uiSampleService.concatenateSequenceFiles(sampleId, sequenceObjectIds, newFileName,
 					removeOriginals));
 		} catch (ConcatenateException e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(null);
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}